move record_to_param_values to arrow_tools crate
